requirements.txt
google
python-dotenv
google.genai